support-intellegence/backend/requirements.txt
# CSIP Backend Dependencies
# Python 3.12+

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
websockets==13.1

# Async HTTP
httpx==0.27.2
aiohttp==3.11.7

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Cache / Message Broker
redis[hiredis]==5.2.0
celery[redis]==5.4.0

# Kafka
aiokafka==0.11.0

# LLM / AI
anthropic==0.39.0
openai==1.57.0  # For embeddings fallback

# Vector Store (local dev: ChromaDB; prod: Pinecone client)
chromadb==0.5.20
pinecone-client==5.0.1

# Pydantic / Settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Security / Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# Secret Management
hvac==2.3.0  # HashiCorp Vault

# PII Masking
presidio-analyzer==2.2.356
presidio-anonymizer==2.2.356
spacy==3.8.3

# Observability
opentelemetry-api==1.28.4
opentelemetry-sdk==1.28.4
opentelemetry-exporter-otlp-proto-grpc==1.28.4
opentelemetry-instrumentation-fastapi==0.49b4
opentelemetry-instrumentation-sqlalchemy==0.49b4
opentelemetry-instrumentation-redis==0.49b4
opentelemetry-instrumentation-httpx==0.49b4

# Circuit Breaker
circuitbreaker==2.0.0

# GitHub Integration
PyGithub==2.5.0
GitPython==3.1.43

# Kubernetes
kubernetes==31.0.0

# AWS
boto3==1.35.80

# Utilities
python-dateutil==2.9.0
pytz==2024.2
structlog==24.4.0
tenacity==9.0.0
xxhash==3.5.0

# PagerDuty
pdpyras==5.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
testcontainers==4.8.2
